*: Fix usage of duplicate cAdvisor time-series
In Kubernetes 1.11 the parent cgroup metrics of all containers in a Pod
also has the `namespace` and `pod_name` labels set. The dashboards
previously summed all time-series that have these labels, however, that
caused primarily memory to be doubled.

As a result, this also removes the non-Pod metrics from the cluster
dashboard.