Update certificate expiration alert message
Update the KubeClientCertificateExpiration alert description to more
accurately reflect the meaning of the metric. This metric is used to
observe the expiration of client certificates used for authentication,
and in this case, for authentication to the apiserver. It does not mean
that the apiserver's server certificate is expiring which can be implied
by the existing message.