Add support for multi-cluster dashboards.
If you are running Thanos (with external labels for clusters)
or Cortex (with a cluster label) use these dashboards to have an
overview of all the clusters in your Kubernetes environment.

The cluster label is configurable and defaults to 'cluster'.
There is a feature toggle, (showMultiCluster) that is off by default.
Enabling multiCluster will show 2 new multiCluster dashboards and
show clusters pickers on all the rest.

Note: you need head of grafana/jsonnet-libs, so jb update

Co-authored-by: Bob Cotton <bob.cotton@gmail.com>