Count overcommit memory only for pods in Pending or Running status
In our cluster we have a bunch of cron jobs. Currently, with the existing query, it appears to be counting the pods that are already completed or failed.

As far as I understood, I believe the overcommit should count only the pods in Pending or Running state.

This change should be applied to the line 64, where it counts the CPU overcommit. It was changed in the PR #18.

I did not change them all because I wan't to check where this is a valid change or not, since it is not very clear to me and my colleagues.