Increase the pods dashboard query rate (#212)
The pods dashboard uses a hardcoded rate interval of 1m in the CPU and Network
I/O panels. A rate interval should be at least four times the scrape interval
in Prometheus, in order to avoid flakes in the graph (see
https://github.com/prometheus/prometheus/issues/4229#issuecomment-395113148).
This means that the scrape interval for the source data would have to be 15s in
order for the graph to display reliably. This commit makes the rate configurable
and so allows the user to select a longer interval if necessary.

* Increase rate interval in pods dashboard to 4m

* Update rate to irate for querties in pods dashboards