Sped up rule query 10x and added "== 0" back
On @brancz's suggestion, I tried to simplify the query by removing the subquery and just using a normal range selector. I had to flip the order of operations a bit in order for this to work, but when I did, average query runtime went from 3000ms to 300ms, and testing confirmed results were the same as the original query down to the 8th decimal place (I assume rounding errors are responsible for the negligible various beyond that).

Also added "== 0" back into the "absent()" operand because I realized that the "code" label leaks into the result if the "== 0" isn't there.