alerts: Alert when a node evicts pods for any reason
Evictions are unusual and may be a sign of resource pressure. While
it is not a strong error signal, it is important to know when these
events occur as they may be symptomatic of workload, node, or cluster
problems. In a healthy cluster, eviction should be rare.

The kubelet_evictions metric was added in Kubernetes 1.16.