Use `default` as a picker value for datasource variable
There can be multiple  datasources of `type: prometheus` available (i.e.
Prometheus and Thanos) and a datasource can be marked as a `default`
which makes any dashboard using the `default` as a datasource
automatically uses the one chosen by the user.

This PR makes the selector on the dashboards use the value of `default`.