dashboards: Default dashboards to UTC
Various tools in the SRE space use UTC as the default to make it
effortless during incidents to not have to calculate between timezones.
Just an example, both Prometheus and Thanos default to this behavior as
well. People can still override this in downstream usage, but I think we
should be promoting best practices.

Signed-off-by: Frederic Branczyk <fbranczyk@gmail.com>