dashboards: Make the API dashboard cluster aware
The dashboard already had a template for multiCluster support but only
some of the queries were using the clusterLabel variable. Unfortunately
the singlestat panels did not work as expected because the clusterLabel
was not being used so multiple values were being returned if multiple
clusters are discoverable. As such, lets also add the clusterLabel
variable to the to the rest of the queries.