Make AggregatedAPIErrors less noisy
This alert has fired routinely for different cloud providers. Without
having done exacts science and having investigated the various SLAs
(should they actually exist for the API), I have concluded that the
expectations of this alerts are too high. The relatively short hiccups
never caused any noticable issues. So this is doubling the time to
average over.

I have also simplified the wording in the details.

`make fmt` created a bunch of reformattings that are unrelated to this
change. I guess that's because of a new version of jsonnet. I can
remove them from this commit if that's undesired here.

Signed-off-by: beorn7 <beorn@grafana.com>