rules/node.libsonnet: fix many-to-many errors for node:node_num_cpu:sum
There's an edge case triggering many-to-many errors for the
node:node_num_cpu:sum rule. After a Prometheus pod with persistent
storage is restarted, there is a short window of time (e.g. 5m) during
which 2 timeseries for the "node_namespace_pod:kube_pod_info:" metric
are returned for the same ("namespace", "pod") labels but with different
"node" label values. This is because Prometheus didn't mark the "old"
series as stale on shutdown.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>