Add a flag for managed K8s clusters
On managed Kubernetes clusters some of the control plane components are not exposed to customers. This adds a flag reflecting that, which removes the alerts/dashboards if enabled.

A feature I'd prefer instead of doing this: https://github.com/prometheus-operator/kube-prometheus/blob/main/jsonnet/kube-prometheus/addons/managed-cluster.libsonnet#L7-L22.