storage: Add ability to mark expected expected 'full volumes' via PVC label
Certain workloads, by design will provision a persistent volume that is
entirely 'full' from the outset. Such an example is KubeVirt CDI
where the file is a disk image, and the underlying PV will be the exact
size of the allocated file. This causes the KubePersistentVolumeFillingUp
storage alerts to fire as soon as they are evaluated, despite the fact
that the data in the volume has reached its maximum size.

This change provides a mechanism which allows cluster admin
to label the PVC in advance and avoid the alerts from ever
firing in these known situations.