Change ordering of workload type and workload in `Compute Resources / Workload`
Prior to this change, `Compute Resources / Workload` dashboard had
`namespace > workload > type` as filters where `type` mearly used as
a label rather a filter.

This commit brings a change in the UX, now the filtering order has changed to
`namespace > type > workload` and all of those filters are actually
functions like a filter.

For example, from now on the user should able to select namespace, then select
the type of workload and finally the actual workload.

Signed-off-by: Arunprasad Rajkumar <arajkuma@redhat.com>