alerts: improve KubeAggregatedAPIErrors
KubeAggregatedAPIErrors was alerting based on the number of errors
returned by all the instances of an aggregated API which made it more
likely to fire the more instances the API had. To improve that we need
to alert at the instance level rather than at the API level by changing
the aggregation function from a `sum` to a `max`.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>