Use `kube_statefulset_replicas` for `KubeStatefulSetReplicasMismatch`
According to [0] and [1] `kube_statefulset_status_replicas` returns
"the number of Pods created by the StatefulSet controller". Therefore
this alert only fires when the StatefulSet controller is able to create
the pod, but it does not get ready.

In cases where the controller is unable to create the pod (e.g. invalid
pod spec) this alert does not fire.
This is contrary to the expectation of firing when the statefulset does
not match the configured amount of replicas.

[0]: https://kubernetes.io/docs/reference/generated/kubernetes-api/v1.32/#statefulsetstatus-v1-apps
[1]: https://github.com/kubernetes/kube-state-metrics/blob/2a95d4649b2fea55799032fb9c0b571c4ba7f776/internal/store/statefulset.go#L75